% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_smooths}
\alias{check_smooths}
\title{Check that only supported smooths are used}
\usage{
check_smooths(form)
}
\arguments{
\item{form}{Model formula}
}
\description{
Currently, this raises an error message if tensor product
smooths are used (te, ti, t2).
}
